+++ pyvoronoi/voronoi.hpp
@@ -333,15 +333,15 @@ public:
 	std::vector<Point> GetPoints();
 	std::vector<Segment> GetSegments();
 
-    long long CountPoints();
-    long long CountSegments();
-	long long CountVertices();
-	long long CountEdges();
-	long long CountCells();
+    unsigned long long CountPoints();
+    unsigned long long CountSegments();
+	unsigned long long CountVertices();
+	unsigned long long CountEdges();
+	unsigned long long CountCells();
 
-    std::vector<long long> GetIntersectingSegments();
-    std::vector<long long> GetDegenerateSegments();
-    std::vector<long long> GetPointsOnSegments();
+    std::vector<unsigned long long> GetIntersectingSegments();
+    std::vector<unsigned long long> GetDegenerateSegments();
+    std::vector<unsigned long long> GetPointsOnSegments();
 
 	//Map index to vertex
 	typedef std::pair<long long, const voronoi_diagram<double>::vertex_type*> index_to_vertex;
@@ -371,11 +371,11 @@ public:
 	void MapEdgeIndexes();
 	void MapCellIndexes();
 
-    Point GetPoint(int index);
-    Segment GetSegment(int index);
-	c_Vertex GetVertex(long long index);
-	c_Edge GetEdge(long long index);
-	c_Cell GetCell(long long index);
+    Point GetPoint(unsigned long long index);
+    Segment GetSegment(unsigned long long index);
+	c_Vertex GetVertex(unsigned long long index);
+	c_Edge GetEdge(unsigned long long index);
+	c_Cell GetCell(unsigned long long index);
 
 
 private:
